Films
This application is made to download movies and learn their information. The application first takes the data with retrofit and then saves it to the local database with room. After that, the movies are shown to the user. The reason for writing to the local database is for the user to see some of the data even if there is no internet